server validation

+++ src/actions/actions.ts
@@ -1,15 +1,23 @@
 "use server";
 
 import prisma from "@/lib/db";
+import { Pet } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet) {
+export async function addPet(pet: unknown) {
     await sleep(2000);
+
+    const validatedPet = petFormSchema.safeParse(pet);
+
+    if (!validatedPet.success) {
+        return { message: "Invalid pet data" };
+    }
 
     try {
     await prisma.pet.create({
-        data: pet,
+        data: validatedPet.data,
     });
 
 } catch (error) {
@@ -18,15 +26,27 @@ export async function addPet(pet) {
 revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: string, newPetData) {
+export async function editPet(petId: unknown, newPetData: unknown) {
 
     sleep(2000);
+
+    const validPetId = petIdSchema.safeParse(petId);
+
+    if (!validPetId.success) {
+        return { message: "Invalid pet ID" };
+    }
+
+    const validatedNewPetData = petFormSchema.safeParse(newPetData);
+
+    if (!validatedNewPetData.success) {
+        return { message: "Invalid pet data" };
+    }
 
     try{
     await prisma.pet.update({
         where: {
-            id: petId },
-            data: newPetData,
+            id: validPetId.data },
+            data: validatedNewPetData.data,
     });
     
     } catch (error) {
@@ -37,13 +57,19 @@ export async function editPet(petId: string, newPetData) {
 
 }
 
-export async function deletePet(petId: string) {
+export async function deletePet(petId: unknown) {
     await sleep(2000);
+
+    const validPetId = petIdSchema.safeParse(petId);
+
+    if (!validPetId.success) {
+        return { message: "Invalid pet ID" };
+    }
     
     try {
     await prisma.pet.delete({
         where: {
-            id: petId
+            id: validPetId.data
         }
     });
